refactor(api): read a stopped merge tail once for both operator reentry verbs
The repair and rerun reentry verbs each re-derived "this Regression Task's
tail is stopped and waiting for an operator" from the same ladder over the
latest recovery attempt, the three bound Tasks, their active Runs and the
recovery Run's verdict, and each implemented request-id idempotency by
scanning operator activity and validating its metadata by hand.

`merge-tail-reentry.ts` states that ladder once. `readStoppedTail` answers it
and returns the aggregate, the recovery context, the Regression Task, the
recovery Run and its verdict; the refusal each rung reports is a per-verb
vocabulary table, so both verbs keep their exact refusal codes and messages.
`priorOperatorRequest` takes the identity fields a verb records and the
settlement that confirms them, so the `typeof` validation and the
`operatorNote` skip live once. The two verbs are now settlements that add
only their own rungs.

Deletes `merge-tail-repair-reentry.ts`, `merge-tail-rerun-reentry.ts` and
`merge-tail-repair-reentry.test.ts`.
