+++ packages/api/src/routes/tasks.ts
@@ -95,8 +95,7 @@ import { jsonValue } from "../execution.js";
 import { refusalFor, type Refusal } from "../refusal.js";
 import { activityInput } from "../run-lifecycle.js";
 import { OPERATOR_NOTE_METADATA_FIELD } from "../run-claim.js";
-import { requestMergeTailRepair } from "../merge-tail-repair-reentry.js";
-import { requestMergeTailRerun } from "../merge-tail-rerun-reentry.js";
+import { requestMergeTailRepair, requestMergeTailRerun } from "../merge-tail-reentry.js";
 import { computeNextOccurrence, validateSchedule } from "../scheduler.js";
 import { patchTask, taskInput, taskPatch } from "../task-patch.js";
 import { lockTaskMutationRows } from "../task-write.js";
